+++ DataMaps/DataMaps/JunimoHutMap.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Microsoft.Xna.Framework;
 using Pathoschild.Stardew.Common;
 using Pathoschild.Stardew.DataMaps.Framework;
-using Pathoschild.Stardew.DataMaps.Framework.Integrations;
+using Pathoschild.Stardew.DataMaps.Framework.Integrations.PelicanFiber;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Buildings;
@@ -28,7 +28,7 @@ namespace Pathoschild.Stardew.DataMaps.DataMaps
         /// <summary>The maximum number of tiles from the center a Junimo hut can harvest.</summary>
         private readonly int MaxRadius = JunimoHut.cropHarvestRadius;
 
-        /// <summary>Handles the logic for integrating with the Pelican Fiber mod.</summary>
+        /// <summary>Handles access to the Pelican Fiber mod.</summary>
         private readonly PelicanFiberIntegration PelicanFiber;
 
 
@@ -47,7 +47,7 @@ namespace Pathoschild.Stardew.DataMaps.DataMaps
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="translations">Provides translations in stored in the mod folder's i18n folder.</param>
-        /// <param name="pelicanFiber">Handles the logic for integrating with the Pelican Fiber mod.</param>
+        /// <param name="pelicanFiber">Handles access to the Pelican Fiber mod.</param>
         public JunimoHutMap(ITranslationHelper translations, PelicanFiberIntegration pelicanFiber)
         {
             this.PelicanFiber = pelicanFiber;