+++ DataMaps/Properties/AssemblyInfo.cs
@@ -1,9 +1,7 @@
-﻿using System.Reflection;
+using System.Reflection;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Pathoschild.Stardew.DataMaps")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyCopyright("Copyright © 2016")]
 [assembly: Guid("cdeef8af-15cb-4955-8163-e446f212c2f4")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+[assembly: AssemblyVersion("1.1.0")]
+[assembly: AssemblyFileVersion("1.1.0")]