+++ ChestsAnywhere/release-notes.md
@@ -1,6 +1,12 @@
 [← back to readme](README.md)
 
 # Release notes
+## 1.12.3
+* Added option to disable remote chest access from certain locations.
+* Fixed shipping bin UI not allowing more than 36 items.
+* Fixed shipping bin UI resetting gamepad cursor position on refresh.
+* Fixed error when the range is set to `None` and a chest is opened directly.
+
 ## 1.12.2
 * Fixed broken shipping bin UI.
 * Fixed deprecated API usage.