+++ DataMaps/release-notes.md
@@ -1,5 +1,13 @@
 # Release notes
+### 1.1
+* Added bee house coverage map.
+* Added support for Cobalt's sprinkler.
+* Added support for Simple Sprinkler's custom sprinkler range.
+* Updated Better Sprinklers support.
+* Fixed deprecated API usage.
+
 ### 1.0
 * Initial version.
-* Added Junimo huts, scarecrows, sprinklers, and accessibility data maps.
-* Added compatibility with Better Sprinklers and Pelican Fiber.
+* Added Junimo hut coverage, scarecrow coverage, sprinkler coverage, and accessibility maps.
+* Added support for Better Sprinklers' custom sprinkler range.
+* Added support for Pelican Fiber's custom build menu.